Deduplicate `Selector`
