+++ packages/graph/hash_graph/lib/graph/src/ontology/mod.rs
@@ -7,9 +7,14 @@ mod link_type;
 mod property_type;
 
 use core::fmt;
+use std::result::Result as StdResult;
 
 use error_stack::{Context, IntoReport, Result, ResultExt};
-use serde::{Deserialize, Serialize, Serializer};
+use serde::{
+    de,
+    de::{Unexpected, Visitor},
+    Deserialize, Deserializer, Serialize, Serializer,
+};
 use serde_json;
 use tokio_postgres::types::{FromSql, ToSql};
 use type_system::{uri::VersionedUri, DataType, EntityType, LinkType, PropertyType};
@@ -20,6 +25,43 @@ pub use self::{
     data_type::DataTypeQueryPath, entity_type::EntityTypeQueryPath, link_type::LinkTypeQueryPath,
     property_type::PropertyTypeQueryPath,
 };
+
+pub enum Selector {
+    Asterisk,
+}
+
+impl<'de> Deserialize<'de> for Selector {
+    fn deserialize<D>(deserializer: D) -> StdResult<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        struct SelectorVisitor;
+
+        impl<'de> Visitor<'de> for SelectorVisitor {
+            type Value = Selector;
+
+            fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
+                formatter.write_str("a wildcard (*)")
+            }
+
+            fn visit_str<E: de::Error>(self, v: &str) -> StdResult<Self::Value, E> {
+                match v {
+                    "*" => Ok(Selector::Asterisk),
+                    _ => Err(de::Error::invalid_value(Unexpected::Str(v), &self)),
+                }
+            }
+
+            fn visit_bytes<E: de::Error>(self, v: &[u8]) -> StdResult<Self::Value, E> {
+                match core::str::from_utf8(v) {
+                    Ok(s) => self.visit_str(s),
+                    Err(_) => Err(E::invalid_value(de::Unexpected::Bytes(v), &self)),
+                }
+            }
+        }
+
+        deserializer.deserialize_str(SelectorVisitor)
+    }
+}
 
 // TODO - find a good place for AccountId, perhaps it will become redundant in a future design
