+++ packages/graph/hash_graph/lib/graph/src/ontology/property_type.rs
@@ -2,13 +2,13 @@ use std::fmt;
 
 use error_stack::{IntoReport, Report};
 use serde::{
-    de::{self, Deserializer, SeqAccess, Unexpected, Visitor},
+    de::{self, Deserializer, SeqAccess, Visitor},
     Deserialize,
 };
 use type_system::PropertyType;
 
 use crate::{
-    ontology::{data_type::DataTypeQueryPathVisitor, DataTypeQueryPath},
+    ontology::{data_type::DataTypeQueryPathVisitor, DataTypeQueryPath, Selector},
     store::query::{Path, QueryRecord},
 };
 
@@ -120,43 +120,6 @@ impl<'de: 'k, 'k> Deserialize<'de> for PropertyTypeQueryPath<'k> {
         D: Deserializer<'de>,
     {
         deserializer.deserialize_seq(PropertyTypeQueryPathVisitor::new(0))
-    }
-}
-
-pub enum Selector {
-    Asterisk,
-}
-
-impl<'de> Deserialize<'de> for Selector {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        struct SelectorVisitor;
-
-        impl<'de> Visitor<'de> for SelectorVisitor {
-            type Value = Selector;
-
-            fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-                formatter.write_str("a wildcard (*)")
-            }
-
-            fn visit_str<E: de::Error>(self, v: &str) -> Result<Self::Value, E> {
-                match v {
-                    "*" => Ok(Selector::Asterisk),
-                    _ => Err(de::Error::invalid_value(Unexpected::Str(v), &self)),
-                }
-            }
-
-            fn visit_bytes<E: de::Error>(self, v: &[u8]) -> Result<Self::Value, E> {
-                match core::str::from_utf8(v) {
-                    Ok(s) => self.visit_str(s),
-                    Err(_) => Err(E::invalid_value(de::Unexpected::Bytes(v), &self)),
-                }
-            }
-        }
-
-        deserializer.deserialize_str(SelectorVisitor)
     }
 }
